Intensity is designed to measure how long sudden jams last. The script to run classify.py also runs it.
The window for rolling apply uses the actual window values provided.

+++ src/characterise/classify.sh
@@ -1,6 +1,10 @@
 #!/bin/bash
 
-ini=$NYCTRAFFICLOG/characterise/classify
+if [ $# -eq 0 ]; then
+    exit 1
+fi
+
+ini=$NYCTRAFFICLOG/characterise/$1
 rm --recursive --force $ini
 mkdir --parents $ini
 
@@ -13,5 +17,5 @@ python3 $NYCTRAFFIC/src/prediction/mkconfigs.py \
 
 for i in $ini/*; do
     echo $i
-    python3 $NYCTRAFFIC/src/characterise/classify.py --config $i || break
+    python3 $NYCTRAFFIC/src/characterise/$1.py --config $i || break
 done
